file refactoring + tweaks to randomess of output
