+++ Cargo.toml
@@ -10,7 +10,7 @@ authors = ["fruit-bird"]
 anyhow = "1.0.69"
 # arboard = "3.2.0"
 clap = { version = "4.1.6", features = ["derive"] }
-google_translator = { git = "https://github.com/fruit-bird/Eveheeero-Translate/" }
+google_translator = { git = "https://github.com/fruit-bird/Eveheeero-Translate/", version = "0.2.3" }
 rand = "0.8.5"
 reqwest = "0.11.14"
 thiserror = "1.0.40"
