+++ src/main.rs
@@ -1,6 +1,6 @@
 // use arboard::Clipboard;
 use clap::Parser;
-use hollow::Prompt;
+use hollow::HollowPrompt;
 
 /// SeEk TRuth
 #[derive(Debug, Parser)]
@@ -12,17 +12,18 @@ struct HollowArgs {
     #[arg(default_value_t = String::from("Moon landing conspiracy"))]
     second_topic: String,
     /// Language to mix into the output
-    #[arg(short, long = "lang", default_value_t = String::from("ja"))]
+    #[arg(short = 'l', long = "lang", default_value_t = String::from("ja"))]
     second_language: String,
     // /// Copy the output to the clipboard
     // #[arg(short, long)]
     // clipboard: bool,
 }
 
+// TODO: change this to a git branch. 2 commits earlier in master is the older, better working version
 #[tokio::main]
 async fn main() {
     let args = HollowArgs::parse();
-    let prompt = Prompt::new(&args.first_topic, &args.second_topic, &args.second_language);
+    let prompt = HollowPrompt::new(&args.first_topic, &args.second_topic, &args.second_language);
 
     let the_spooky = match prompt.run().await {
         Ok(entry) => entry,
